Move event creation methods into `EventFactory`
