+++ client/src/test/java/org/spine3/base/EventPredicatesShould.java
@@ -24,6 +24,7 @@ import com.google.common.base.Predicate;
 import com.google.common.testing.NullPointerTester;
 import com.google.protobuf.Timestamp;
 import org.junit.Test;
+import org.spine3.server.command.EventFactory;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
@@ -43,8 +44,8 @@ import static org.spine3.test.TimeTests.Past.secondsAgo;
 public class EventPredicatesShould {
 
     private static Event createEvent(int minutesAgo) {
-        final Event result = Events.createEvent(newUuidValue(),
-                                                newEventContext(minutesAgo(minutesAgo)));
+        final Event result = EventFactory.createEvent(newUuidValue(),
+                                                      newEventContext(minutesAgo(minutesAgo)));
         return result;
     }
