+++ server/src/main/java/org/spine3/server/aggregate/Aggregate.java
@@ -42,9 +42,9 @@ import java.lang.reflect.InvocationTargetException;
 import java.util.List;
 import java.util.Set;
 
-import static org.spine3.base.Events.createEvent;
 import static org.spine3.base.Events.getMessage;
 import static org.spine3.protobuf.Timestamps2.getCurrentTime;
+import static org.spine3.server.command.EventFactory.createEvent;
 import static org.spine3.server.reflect.EventApplierMethod.forEventMessage;
 import static org.spine3.util.Exceptions.wrappedCause;
 import static org.spine3.validate.Validate.isNotDefault;
