+++ server/src/main/java/org/spine3/server/command/EventFactory.java
@@ -1,0 +1,135 @@
+/*
+ * Copyright 2017, TeamDev Ltd. All rights reserved.
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package org.spine3.server.command;
+
+import com.google.protobuf.Any;
+import com.google.protobuf.Message;
+import com.google.protobuf.Timestamp;
+import org.spine3.base.CommandContext;
+import org.spine3.base.Event;
+import org.spine3.base.EventContext;
+import org.spine3.base.EventId;
+import org.spine3.base.Identifiers;
+import org.spine3.base.Version;
+
+import javax.annotation.Nullable;
+
+import static com.google.common.base.Preconditions.checkNotNull;
+import static org.spine3.protobuf.Messages.toAny;
+import static org.spine3.protobuf.Timestamps2.getCurrentTime;
+
+/**
+ * Produces events in response to a command.
+ *
+ * @author Alexander Yevsyukov
+ */
+public class EventFactory {
+
+    private EventFactory() {}
+
+    /** Generates a new random UUID-based {@code EventId}. */
+    public static EventId generateId() {
+        final String value = Identifiers.newUuid();
+        return EventId.newBuilder()
+                      .setUuid(value)
+                      .build();
+    }
+
+    /**
+     * Creates a new {@code Event} instance.
+     *
+     * @param messageOrAny the event message or {@code Any} containing the message
+     * @param context      the event context
+     * @return created event instance
+     */
+    public static Event createEvent(Message messageOrAny, EventContext context) {
+        checkNotNull(messageOrAny);
+        checkNotNull(context);
+        final Any packed = toAny(messageOrAny);
+        final Event result = Event.newBuilder()
+                                  .setMessage(packed)
+                                  .setContext(context)
+                                  .build();
+        return result;
+    }
+
+    /**
+     * Creates {@code Event} instance for import or integration operations.
+     *
+     * @param event      the event message
+     * @param producerId the ID of an entity which is generating the event packed into {@code Any}
+     * @return event with data from an external source
+     */
+    public static Event createImportEvent(Message event, Any producerId) {
+        checkNotNull(event);
+        checkNotNull(producerId);
+        final EventContext context = createImportEventContext(producerId);
+        final Event result = createEvent(event, context);
+        return result;
+    }
+
+    /**
+     * Creates {@code EventContext} instance for an event generated during data import.
+     *
+     * <p>The method does not set {@code CommandContext} because there was no command.
+     *
+     * <p>The {@code version} attribute is not populated either.
+     * It is the responsibility of the target aggregate to populate the missing fields.
+     *
+     * @param producerId the ID of the producer which generates the event packed into {@code Any}
+     * @return new instance of {@code EventContext} for the imported event
+     */
+    public static EventContext createImportEventContext(Any producerId) {
+        checkNotNull(producerId);
+        final EventContext.Builder builder = EventContext.newBuilder()
+                                                         .setEventId(generateId())
+                                                         .setTimestamp(getCurrentTime())
+                                                         .setProducerId(producerId);
+        return builder.build();
+    }
+
+    /**
+     * Creates new {@code CommandContext} with passed parameters.
+     *
+     * @param producerId the ID of an object which produced the event
+     * @param version optional version of the object
+     * @param commandContext the context of the command handling of which produced the event
+     * @return new {@code CommandContext}
+     */
+    public static EventContext createEventContext(Any producerId,
+                                                  @Nullable Version version,
+                                                  CommandContext commandContext) {
+        checkNotNull(producerId);
+        checkNotNull(commandContext);
+
+        final EventId eventId = generateId();
+        final Timestamp timestamp = getCurrentTime();
+        final EventContext.Builder builder = EventContext.newBuilder()
+                                                         .setEventId(eventId)
+                                                         .setTimestamp(timestamp)
+                                                         .setCommandContext(commandContext)
+                                                         .setProducerId(producerId);
+        if (version != null) {
+            builder.setVersion(version);
+        }
+        return builder.build();
+    }
+}
