+++ server/src/test/java/org/spine3/server/event/Given.java
@@ -28,10 +28,10 @@ import org.spine3.base.CommandContext;
 import org.spine3.base.Event;
 import org.spine3.base.EventContext;
 import org.spine3.base.EventId;
-import org.spine3.base.Events;
 import org.spine3.base.Identifiers;
 import org.spine3.people.PersonName;
 import org.spine3.protobuf.AnyPacker;
+import org.spine3.server.command.EventFactory;
 import org.spine3.server.event.enrich.EventEnricher;
 import org.spine3.test.Tests;
 import org.spine3.test.event.ProjectCompleted;
@@ -49,8 +49,8 @@ import org.spine3.users.UserId;
 
 import javax.annotation.Nullable;
 
-import static org.spine3.base.Events.createEvent;
 import static org.spine3.base.Identifiers.newUuid;
+import static org.spine3.server.command.EventFactory.createEvent;
 import static org.spine3.test.EventTests.newEvent;
 import static org.spine3.test.Tests.newTenantUuid;
 import static org.spine3.testdata.TestEventContextFactory.createEventContext;
@@ -195,7 +195,7 @@ public class Given {
         private static Event createGenericEvent(Message eventMessage) {
             final Any wrappedMessage = AnyPacker.pack(eventMessage);
             final EventContext eventContext = createEventContext();
-            final Event permissionRevoked = Events.createEvent(wrappedMessage, eventContext);
+            final Event permissionRevoked = EventFactory.createEvent(wrappedMessage, eventContext);
             return permissionRevoked;
         }
     }
