+++ server/src/test/java/org/spine3/server/projection/ProjectionRepositoryShould.java
@@ -39,6 +39,7 @@ import org.spine3.protobuf.AnyPacker;
 import org.spine3.protobuf.Durations2;
 import org.spine3.protobuf.Timestamps2;
 import org.spine3.server.BoundedContext;
+import org.spine3.server.command.EventFactory;
 import org.spine3.server.entity.RecordBasedRepository;
 import org.spine3.server.entity.RecordBasedRepositoryShould;
 import org.spine3.server.entity.idfunc.IdSetEventFunction;
@@ -201,7 +202,7 @@ public class ProjectionRepositoryShould
     }
 
     private void checkDispatchesEvent(Message eventMessage) {
-        final Event event = Events.createEvent(eventMessage, createEventContext(ID, tenantId()));
+        final Event event = EventFactory.createEvent(eventMessage, createEventContext(ID, tenantId()));
 
         keepTenantIdFromEvent(event);
 
@@ -225,7 +226,7 @@ public class ProjectionRepositoryShould
     private void checkDoesNotDispatchEventWith(Status status) {
         repository().setStatus(status);
         final ProjectCreated eventMsg = projectCreated();
-        final Event event = Events.createEvent(eventMsg, createEventContext(ID, tenantId()));
+        final Event event = EventFactory.createEvent(eventMsg, createEventContext(ID, tenantId()));
 
         repository().dispatch(event);
 
@@ -236,8 +237,8 @@ public class ProjectionRepositoryShould
     public void throw_exception_if_dispatch_unknown_event() {
         final StringValue unknownEventMessage = StringValue.getDefaultInstance();
 
-        final Event event = Events.createEvent(unknownEventMessage,
-                                               EventContext.getDefaultInstance());
+        final Event event = EventFactory.createEvent(unknownEventMessage,
+                                                     EventContext.getDefaultInstance());
 
         repository().dispatch(event);
     }
@@ -385,7 +386,7 @@ public class ProjectionRepositoryShould
                                                              .setTenantId(tenantId()))
                             .setTimestamp(Timestamps2.getCurrentTime())
                             .build();
-        final Event event = Events.createEvent(eventMessage, context);
+        final Event event = EventFactory.createEvent(eventMessage, context);
         appendEvent(boundedContext.getEventBus()
                                   .getEventStore(), event);
         // Set up repository
@@ -422,7 +423,7 @@ public class ProjectionRepositoryShould
                                                      .setProducerId(AnyPacker.pack(projectId))
                                                      .setTimestamp(Timestamps2.getCurrentTime())
                                                      .build();
-            final Event event = Events.createEvent(eventMessage, context);
+            final Event event = EventFactory.createEvent(eventMessage, context);
             appendEvent(eventStore, event);
         }
         // Set up repository
