+++ server/src/test/java/org/spine3/testdata/Sample.java
@@ -34,8 +34,8 @@ import org.spine3.base.CommandContext;
 import org.spine3.base.Commands;
 import org.spine3.base.Event;
 import org.spine3.base.EventContext;
-import org.spine3.base.Events;
 import org.spine3.protobuf.AnyPacker;
+import org.spine3.server.command.EventFactory;
 import org.spine3.type.TypeUrl;
 import org.spine3.users.TenantId;
 
@@ -65,14 +65,14 @@ public class Sample {
                                 TenantId tenantId) {
         final EventContext eventContext = TestEventContextFactory.createEventContext(producerId, tenantId);
         final Message eventMessage = messageOfType(eventClass);
-        final Event event = Events.createEvent(eventMessage, eventContext);
+        final Event event = EventFactory.createEvent(eventMessage, eventContext);
         return event;
     }
 
     public static Event eventBy(Message producerId, Message eventMessage,
                                 TenantId tenantId) {
         final EventContext eventContext = TestEventContextFactory.createEventContext(producerId, tenantId);
-        final Event event = Events.createEvent(eventMessage, eventContext);
+        final Event event = EventFactory.createEvent(eventMessage, eventContext);
         return event;
     }
