+++ server/src/test/java/org/spine3/testdata/TestEventContextFactory.java
@@ -27,7 +27,7 @@ import org.spine3.base.CommandContext;
 import org.spine3.base.Enrichment;
 import org.spine3.base.EventContext;
 import org.spine3.base.EventId;
-import org.spine3.base.Events;
+import org.spine3.server.command.EventFactory;
 import org.spine3.server.integration.IntegrationEventContext;
 import org.spine3.test.TestCommandFactory;
 import org.spine3.users.TenantId;
@@ -60,7 +60,7 @@ public class TestEventContextFactory {
         final Timestamp now = getCurrentTime();
         final UserId userId = newUserId(newUuid());
         final CommandContext commandContext = createCommandContext(userId, generateId(), now);
-        final EventId eventId = Events.generateId();
+        final EventId eventId = EventFactory.generateId();
         final EventContext.Builder builder = EventContext.newBuilder()
                                                          .setEventId(eventId)
                                                          .setCommandContext(commandContext)
@@ -71,7 +71,7 @@ public class TestEventContextFactory {
 
     public static EventContext createEventContext(Message aggregateId,
                                                   TenantId tenantId) {
-        final EventId eventId = Events.generateId();
+        final EventId eventId = EventFactory.generateId();
         final CommandContext commandContext =
                 TestCommandFactory.newInstance(TestEventContextFactory.class, tenantId)
                                   .createContext();
@@ -94,7 +94,7 @@ public class TestEventContextFactory {
 
     /** Creates a new {@link IntegrationEventContext} with default properties. */
     public static IntegrationEventContext createIntegrationEventContext() {
-        final EventId eventId = Events.generateId();
+        final EventId eventId = EventFactory.generateId();
         final IntegrationEventContext.Builder builder =
                 IntegrationEventContext.newBuilder()
                                        .setEventId(eventId)
@@ -105,7 +105,7 @@ public class TestEventContextFactory {
     }
 
     public static IntegrationEventContext createIntegrationEventContext(Message aggregateId) {
-        final EventId eventId = Events.generateId();
+        final EventId eventId = EventFactory.generateId();
         final IntegrationEventContext.Builder builder = IntegrationEventContext.newBuilder()
                                                                                .setEventId(eventId)
                                                                                .setTimestamp(getCurrentTime())
@@ -116,7 +116,7 @@ public class TestEventContextFactory {
     }
 
     public static EventContext createEventContext(Timestamp timestamp) {
-        final EventId eventId = Events.generateId();
+        final EventId eventId = EventFactory.generateId();
         final EventContext.Builder builder = EventContext.newBuilder()
                                                          .setEventId(eventId)
                                                          .setTimestamp(timestamp)
@@ -125,7 +125,7 @@ public class TestEventContextFactory {
     }
 
     public static EventContext createEventContext(Message aggregateId, Timestamp timestamp) {
-        final EventId eventId = Events.generateId();
+        final EventId eventId = EventFactory.generateId();
         final EventContext.Builder builder = EventContext.newBuilder()
                                                          .setEventId(eventId)
                                                          .setTimestamp(timestamp)
